+++ BookBook.Presentation/AssemblyReference.cs
@@ -1,0 +1,6 @@
+﻿namespace BookBook.Presentation;
+
+public static class AssemblyReference
+{
+
+}
