+++ BookBook.Repository/Implement/AuthorRepositoryV2.cs
@@ -8,4 +8,8 @@ public class AuthorRepositoryV2 : RepositoryBase<Author>, IAuthorRepositoryV2
     public AuthorRepositoryV2(RepositoryContext _repositoryContext) : base(_repositoryContext)
     {
     }
+
+    public IEnumerable<Author> GetAllAuthors(bool trackChanges) 
+            => FindAll(trackChanges).OrderBy(v => v.FirstName + v.LastName).ToList();
+            
 }
