+++ BookBook.Service/IAuthorService.cs
@@ -1,6 +1,9 @@
-﻿namespace BookBook.Service;
+﻿using BookBook.Models.Models;
+
+namespace BookBook.Service;
 
 public interface IAuthorService
 {
+    IEnumerable<Author> GetAllAuthors(bool trackChanges);
 
 }
