+++ BookBook.Service/Implement/AuthorService.cs
@@ -1,4 +1,5 @@
-﻿using Contracts;
+﻿using BookBook.Models.Models;
+using Contracts;
 
 namespace BookBook.Service;
 
@@ -10,5 +11,19 @@ public class AuthorService : IAuthorService
     {
         _repositoryManager = repositoryManager;
         _loggerManager = loggerManager;
+    }
+
+    public IEnumerable<Author> GetAllAuthors(bool trackChanges)
+    {
+        try
+        {
+            var authors = _repositoryManager.AuthorRepositoryV2.GetAllAuthors(trackChanges);
+            return authors;
+        }
+        catch(Exception ex)
+        {
+            _loggerManager.LogError($"Something went wrong in the {nameof(GetAllAuthors)} service method {ex}");
+            throw;
+        }
     }
 }
