+++ BookBook.sln
@@ -17,6 +17,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LoggerService", "LoggerServ
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BookBook.Service", "BookBook.Service\BookBook.Service.csproj", "{42E5B7CE-3BA7-4FEC-AF8E-FD270D365A07}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BookBook.Presentation", "BookBook.Presentation\BookBook.Presentation.csproj", "{87866C61-B025-43AD-BA27-DDD58F1A181F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -54,5 +56,9 @@ Global
 		{42E5B7CE-3BA7-4FEC-AF8E-FD270D365A07}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{42E5B7CE-3BA7-4FEC-AF8E-FD270D365A07}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{42E5B7CE-3BA7-4FEC-AF8E-FD270D365A07}.Release|Any CPU.Build.0 = Release|Any CPU
+		{87866C61-B025-43AD-BA27-DDD58F1A181F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{87866C61-B025-43AD-BA27-DDD58F1A181F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{87866C61-B025-43AD-BA27-DDD58F1A181F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{87866C61-B025-43AD-BA27-DDD58F1A181F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
